move cmap/format_12 out of notofonts directory
