+++ tests/test_checks_cmap.py
@@ -1,0 +1,8 @@
+def NOT_IMPLEMENTED_test_check_cmap_format_12():
+    """Check that format 12 cmap subtables are correctly constituted."""
+    # check = CheckTester("cmap/format_12")
+    # TODO: Implement-me!
+    #
+    # code-paths:
+    # - FAIL, ""
+    # - PASS
